check if vm is not created by monitoring
There are situation when between spawning VM on compute site and
registering this machine inside atmosphere DB VM monitoring is triggered.
As a conclusion VM with duplicated `id_at_site` is create, which causes
`Unable to assign VM to appliance` error. Here we are fixing this issue
by checking if VM exists in database or initializing it
(`find_or_initialize`).