take into account created at when checking vm age
Childhood age is used to fix the problem when following situation occurs:
  - atmosphere send request to start VM
  - after request is sent monitoring is triggered, new VM is discovered
    and registered as managed_by_atmosphere = false
  - atmosphere is trying to save VM created in first step into DB. As a result
    exception is thrown (VM with id_at_site already exists)

We used `server.created` from compute site to fix this problem. It have one
drawback - when compute site host data is different that date on the host where
atmosphere is hosted. To minimize this problem this commit introduces checking
at the beginning `created_at` field and only when it is empty `created` date
returned by the compute site is used.