setting version has precedence over setting AT
It solved the problem when VMT is migrated into Amazon. There it is
not possible to register template and set its metadata as atomic
operation. As a conclusion VMT was first discovered but not assigned to
appliance type and version was not set. Next after metadata was added
monitoring assigns this VMT into appliance type and set version into
`appliance_type.version + 1`. This is obviously bug. This commit
fixes this issue, by allowing to update version by the monitoring when
VMT is young. What is more during "young" period if administrator
changes source VMT version it will be populated into migrated VMT.