Work through (or around) Hound change requests.
This commit contains changes which are designed to get everyone's favorite dog to keep it down for a while. Due to the fact that the 80 character line length limit is notoriously exceeded in the Atmosphere codebase, instead of hunting down each instance of this issue (which would be extremely tedious) the commit resets the length limit to 120 characters.