Refactor VMT and Flavor selection logic to acknowledge Tenant-User relationship through Funds.
The Optimizer has been altered in such a way that it only preselects those Virtual Machine Templates which belong to Tenants accessible to the Appliance's owner (i.e. those which share at least one Fund with the user who owns the Appliance's ApplianceSet).

This change broke a lot of old RSpecs due to the way factories operate (creating superfluous User and Tenant objects which are not bound to one another). Fixes were applied wherever required.