validate that admin provided valid AS type
https://github.com/dice-cyfronet/atmosphere/issues/221

Non-admin users were not allowed by abilities to create appl set with arbitrary type,
however admin user could do it. If admin provided an invalid type a 500 error occured
due to invoking a method on a nil object.
Since this commit it is checked if appliance set type is properly set
and an appropriate error message and 422 status code are returned in response.
Fix #221.